requirements.txt
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
cmake==3.26.3
colorama==0.4.6
cryptography==40.0.2
deprecation==2.1.0
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==6.1.0
firebase-functions==0.1.0
Flask==2.3.2
Flask-Cors==3.0.10
functions-framework==3.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-events==0.6.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.3.4